AB/pyproject.toml
[project]
name = "ab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate==0.24.1",
    "audio-recorder-streamlit>=0.0.10",
    "datasets>=3.6.0",
    "evaluate>=0.4.4",
    "ffmpeg-python>=0.2.0",
    "huggingface-hub>=0.33.2",
    "jiwer>=4.0.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "openai-whisper>=20250625",
    "pandas>=2.3.0",
    "pretty-midi>=0.2.10",
    "pyfluidsynth>=1.3.4",
    "soundfile>=0.13.1",
    "streamlit>=1.46.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "transformers==4.35.2",
    "wandb>=0.21.0",
]
